before cleaning
