+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -23,10 +23,10 @@
         </activity>
         <activity
             android:theme="@style/AppTheme.NoActionBar"
-            android:name=".view.PersonActivity" />
+            android:name=".view.contact.PersonActivity" />
         <activity
             android:theme="@style/AppTheme.NoActionBar"
-            android:name=".view.OrganizationActivity"></activity>
+            android:name=".view.organization.OrganizationActivity"></activity>
     </application>
 
 </manifest>
