--- app/src/main/res/drawable/img.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
-
-    <item android:drawable="@drawable/ic_launcher_background"/>
-    <item android:drawable="@drawable/circle"/>
-
-</layer-list>
