+++ app/src/main/res/layout/activity_main.xml
@@ -4,28 +4,28 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    android:elevation="0dp"
+    tools:context=".view.MainActivity">
 
     <com.ogaclejapan.smarttablayout.SmartTabLayout
         android:id="@+id/viewPagerTab"
+        android:elevation="0dp"
         android:layout_width="match_parent"
         android:layout_height="46dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:stl_clickable="true"
-        app:stl_defaultTabBackground="?attr/selectableItemBackground"
+        android:background="@color/colorPrimary"
         app:stl_defaultTabTextAllCaps="true"
-        app:stl_defaultTabTextColor="#FC000000"
+        app:stl_defaultTabTextColor="@color/colorPrimaryBlank"
         app:stl_defaultTabTextHorizontalPadding="16dp"
         app:stl_defaultTabTextMinWidth="0dp"
         app:stl_defaultTabTextSize="12sp"
-        app:stl_distributeEvenly="false"
-        app:stl_dividerColor="#4D000000"
-        app:stl_dividerThickness="1dp"
+        app:stl_distributeEvenly="true"
         app:stl_drawDecorationAfterTab="false"
         app:stl_indicatorAlwaysInCenter="false"
-        app:stl_indicatorColor="#40C4FF"
+        app:stl_indicatorColor="@color/coloSelect"
         app:stl_indicatorCornerRadius="2dp"
         app:stl_indicatorGravity="bottom"
         app:stl_indicatorInFront="false"
@@ -33,10 +33,8 @@
         app:stl_indicatorThickness="4dp"
         app:stl_indicatorWidth="auto"
         app:stl_indicatorWithoutPadding="false"
-        app:stl_overlineColor="#4D000000"
         app:stl_overlineThickness="0dp"
         app:stl_titleOffset="24dp"
-        app:stl_underlineColor="#4D000000"
         app:stl_underlineThickness="1dp" />
 
     <android.support.v4.view.ViewPager
