+++ app/src/main/res/layout/contacts_fragment_list.xml
@@ -4,12 +4,12 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".view.ContactsListFragment">
+    tools:context=".view.contact.ContactsListFragment">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
-
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:background="@color/colorPrimary">
 
         <android.support.v7.widget.RecyclerView
             android:id="@+id/recyclerView"
