+++ app/src/main/res/layout/organizations_list_fragment.xml
@@ -5,7 +5,8 @@
     android:layout_width="match_parent"
 
     android:layout_height="match_parent"
-    tools:context=".view.OrganizationsListFragment">
+    tools:context=".view.organization.OrganizationsListFragment"
+    android:background="@color/colorPrimary">
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
