+++ app/src/main/res/values-v21/styles.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <style name="MyToolbarStyle" parent="Base.Widget.AppCompat.Toolbar">
+        <item name="android:elevation">4dp</item>
+        <item name="android:background">@color/colorPrimary</item>
+    </style>
+</resources>
