+++ app/src/main/res/values/colors.xml
@@ -1,9 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
+    <color name="colorText">#545454</color>
+    <color name="colorPrimary">#2e4159</color>
+    <color name="colorPrimaryDark">#263238</color>
     <color name="colorAccent">#FF4081</color>
     <color name="colorPrimaryBlank">#ffffff</color>
     <color name="colorSecundaryBlank">#5fb6f9</color>
+    <color name="coloSelect">#ffc000</color>
     <color name="colorTransparentColor">#80000000</color>
 </resources>
