+++ app/src/main/res/values/styles.xml
@@ -13,4 +13,13 @@
         <item name="windowNoTitle">true</item>
         <item name="android:windowFullscreen">true</item>
     </style>
+
+    <style name="MyToolbarStyle" parent="Base.Widget.AppCompat.Toolbar">
+        <!--<item name="android:elevation">4dp</item>-->
+        <item name="android:background">@color/colorPrimary</item>
+    </style>
+
+    <style name="MyToolbarTheme" parent="Base.Widget.AppCompat.Toolbar">
+        <item name="android:textColorPrimary">#ffffff</item>
+    </style>
 </resources>
